# expworld/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Swagger Movie - OpenAPI 3.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.17"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "tags": [
        {
            "name": "query",
            "description": "API queries based on city and optional country code."
        }
    ],
    "paths": {
        "/user/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh user token",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "example": "ajsonwebtoken"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token refreshed"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "error": {
                                        "type": "string",
                                        "example": "false"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Token invalidated"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "description": "Login a user with email and password",
                "operationId": "loginUser",
                "consumes": [
                    "application/json"
                ],
                "requestBody": {
                    "description": "User registration parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "example": {
                                    "email": "user@example.com",
                                    "password": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{email}/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile by email",
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "schema": {
                            "type": "string",
                            "format": "email"
                        },
                        "required": true,
                        "description": "Email address of the user"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "mike@gmail.com"
                                        },
                                        "firstName": {
                                            "type": "string",
                                            "example": "Michael"
                                        },
                                        "lastName": {
                                            "type": "string",
                                            "example": "Jordan"
                                        },
                                        "dob": {
                                            "type": "string",
                                            "example": "1990-01-01"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "123 Street, City"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user profile by email",
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "schema": {
                            "type": "string",
                            "format": "email"
                        },
                        "required": true,
                        "description": "Email address of the user"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "Updated user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "firstName": {
                                            "type": "string"
                                        },
                                        "lastName": {
                                            "type": "string"
                                        },
                                        "dob": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration",
                "description": "Register a new user with email and password",
                "operationId": "registerUser",
                "consumes": [
                    "application/json"
                ],
                "requestBody": {
                    "description": "User registration parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "dob": {
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "example": {
                                    "email": "user@example.com",
                                    "firstName": "Shawn",
                                    "lastName": "Mike",
                                    "dob": "2022-5-20",
                                    "address": "fake street No.1",
                                    "password": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful registration",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/data/{imdbID}": {
            "get": {
                "summary": "Get movie data by IMDb ID",
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "name": "imdbID",
                        "in": "path",
                        "description": "IMDb ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "title": "Star Trek: First Contact",
                                    "year": 1996,
                                    "runtime": 111,
                                    "genres": [
                                        "Action",
                                        "Adventure",
                                        "Drama"
                                    ],
                                    "country": "United States",
                                    "principals": [
                                        {
                                            "id": "nm0005772",
                                            "category": "cinematographer",
                                            "name": "Matthew F. Leonetti",
                                            "characters": []
                                        },
                                        {
                                            "id": "nm0001772",
                                            "category": "actor",
                                            "name": "Patrick Stewart",
                                            "characters": [
                                                "Picard"
                                            ]
                                        },
                                        {
                                            "id": "nm0000408",
                                            "category": "actor",
                                            "name": "Jonathan Frakes",
                                            "characters": [
                                                "Riker"
                                            ]
                                        },
                                        {
                                            "id": "nm0000653",
                                            "category": "actor",
                                            "name": "Brent Spiner",
                                            "characters": [
                                                "Data"
                                            ]
                                        },
                                        {
                                            "id": "nm0000996",
                                            "category": "actor",
                                            "name": "LeVar Burton",
                                            "characters": [
                                                "Geordi"
                                            ]
                                        },
                                        {
                                            "id": "nm0734472",
                                            "category": "writer",
                                            "name": "Gene Roddenberry",
                                            "characters": []
                                        },
                                        {
                                            "id": "nm0075834",
                                            "category": "writer",
                                            "name": "Rick Berman",
                                            "characters": []
                                        },
                                        {
                                            "id": "nm0103804",
                                            "category": "writer",
                                            "name": "Brannon Braga",
                                            "characters": []
                                        },
                                        {
                                            "id": "nm0601822",
                                            "category": "writer",
                                            "name": "Ronald D. Moore",
                                            "characters": []
                                        },
                                        {
                                            "id": "nm0000025",
                                            "category": "composer",
                                            "name": "Jerry Goldsmith",
                                            "characters": []
                                        }
                                    ],
                                    "ratings": [
                                        {
                                            "source": "Internet Movie Database",
                                            "value": 7.6
                                        }
                                    ],
                                    "boxoffice": 92027888,
                                    "poster": "https://m.media-amazon.com/images/M/MV5BYzMzZmE3MTItODYzYy00YWI5LWFkNWMtZTY5NmU2MDkxYWI1XkEyXkFqcGdeQXVyMjUzOTY1NTc@._V1_SX300.jpg",
                                    "plot": "The Borg travel back in time intent on preventing Earth's first contact with an alien species. Captain Picard and his crew pursue them to ensure that Zefram Cochrane makes his maiden flight reaching warp speed."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": true,
                                    "message": "Invalid IMDb ID"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": true,
                                    "message": "Movie not found"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/search": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Search for movies",
                "parameters": [
                    {
                        "name": "primaryTitle",
                        "in": "query",
                        "description": "Search keyword for movie titles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Search movie year",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Filter the movies using page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": [
                                        {
                                            "title": "Rush",
                                            "year": 1991,
                                            "imdbID": "tt0102820",
                                            "imdbRating": 6.7,
                                            "rottenTomatoesRating": 71,
                                            "metacriticRating": 66,
                                            "classification": "R"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": true,
                                    "message": "Invalid request"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": true,
                                    "message": "Error in MySQL query"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/people/{id}": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get details of a person",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the person",
                        "required": true,
                        "type": "string"
                    }
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "name": "Patrick Stewart",
                                    "birthYear": 1940,
                                    "deathYear": null,
                                    "roles": [
                                        {
                                            "movieName": "Star Trek: First Contact",
                                            "movieId": "tt0117731",
                                            "category": "actor",
                                            "characters": [
                                                "Picard"
                                            ],
                                            "imdbRating": 7.6
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": true,
                                    "message": "Invalid token"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": true,
                                    "message": "Person not found"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": true,
                                    "message": "Error in MySQL query"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": [
                    "http",
                    "https"
                ],
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}